requirements.txt
fastapi
uvicorn
gtts
python-multipart
google-generativeai
